🚚[Rename] 이미지 필드 분리
