+++ src/features/club-create/container/index.ts
@@ -1,2 +1,3 @@
 export { default as FormContainer } from './FormContainer';
 export { default as CreateBookClubPage } from './CreateBookClubPage';
+export { default as ImageField } from './ImageField';
